A bunch of changes in ManetRoutingBase.
Now it’s possible to specify the name of interfaces when the manet routing protocol is active, see BaseRouting.ned
New parameter manetLabelRouting ,  the routing entries are marked IPRoute::MANET, this label force an exact search with independent of the mask. If this parameter is false the Routing Table will search the first masked route. see BaseRouting.ned
Now it's impossible to call a ManetRoutingBase method without register before (registerRoutingModule)
If the routing mask passed with omnet_chg_rte is ‘0’ the method will use the mask IPAddress::ALLONES_ADDRESS (255.255.255.255)like default address.